+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,56 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+
+    public TxUserService(UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(User user) {
+        userService.save(user);
+    }
+
+    // override 대상인 메서드는 userService의 메서드를 그대로 위임(delegate)한다.
+    @Override
+    public void changePassword(long id, String newPassword, String createdBy) {
+        // 트랜잭션 처리 영역
+        DataSource dataSource = DataSourceConfig.getInstance();
+        Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+            userService.changePassword(id, newPassword, createdBy);
+            connection.commit();
+            connection.setAutoCommit(true);
+        } catch (Exception e) {
+            rollback(e, connection);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+        // 트랜잭션 처리 영역
+    }
+
+    private void rollback(Exception e, Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException ex) {
+            throw new DataAccessException(ex);
+        }
+        throw new DataAccessException(e);
+    }
+}
